+++ kubernetes_state/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - kubernetes_state
 
-## 5.6.3 / 2020-12-23
+## 5.6.3 / 2020-12-23 / Agent 7.25.0
 
 * [Fixed] Fix deployment count metric. See [#8247](https://github.com/DataDog/integrations-core/pull/8247).