+++ mongo/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - mongo
 
-## 2.1.1 / 2020-12-11
+## 2.1.1 / 2020-12-11 / Agent 7.25.0
 
 * [Fixed] Log custom queries which return an empty result set. See [#8105](https://github.com/DataDog/integrations-core/pull/8105).