+++ spark/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * [Added] Add metrics for structured streams. See [#8078](https://github.com/DataDog/integrations-core/pull/8078).
 
-## 1.18.0 / 2020-11-23
+## 1.18.0 / 2020-11-23 / Agent 7.25.0
 
 * [Added] Add more granular executor metrics. See [#8028](https://github.com/DataDog/integrations-core/pull/8028).